Cms

See merge request hardiksukhadiya/faldax-CMS!158

+++ src/containers/Page/Assets/editAssetLimit.js
@@ -152,12 +152,12 @@ class EditableTable extends React.Component {
             let formData = {
                 id: newData[index].id,
                 coin_id: newData[index].coin_id,
-                daily_withdraw_crypto: parseInt(row.daily_withdraw_crypto),
-                daily_withdraw_fiat: parseInt(row.daily_withdraw_fiat),
-                min_withdrawl_crypto: parseInt(row.min_withdrawl_crypto),
-                min_withdrawl_fiat: parseInt(row.min_withdrawl_fiat),
-                monthly_withdraw_crypto: parseInt(row.monthly_withdraw_crypto),
-                monthly_withdraw_fiat: parseInt(row.monthly_withdraw_fiat)
+                daily_withdraw_crypto: parseFloat(row.daily_withdraw_crypto),
+                daily_withdraw_fiat: parseFloat(row.daily_withdraw_fiat),
+                min_withdrawl_crypto: parseFloat(row.min_withdrawl_crypto),
+                min_withdrawl_fiat: parseFloat(row.min_withdrawl_fiat),
+                monthly_withdraw_crypto: parseFloat(row.monthly_withdraw_crypto),
+                monthly_withdraw_fiat: parseFloat(row.monthly_withdraw_fiat)
             }
 
             _this.setState({ loader: true });
